+++ webapp/slide/__init__.py
@@ -1,0 +1,2 @@
+from .generate_questions_from_images import OcrWrapper, GoolabWrapper
+from .slidecapture import SlideCapture, SlideCaptureError
