+++ webapp/slide/googleocr.py
